requirements.txt
Flask
Flask-Login
python-dotenv
email_validator
supabase
